# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Universal Rest API",
        "title": "Unintegrate API",
        "termsOfService": "https://github.com/clysec/dagosy-worker",
        "contact": {
            "name": "Cloudyne Support",
            "email": "support@cloudyne.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/",
    "paths": {
...
        "/api/v1/rest/request": {
            "get": {
                "security": [
                    {
                        "OauthSE": []
                    }
                ],
                "description": "Make a GET Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a GET Request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "OauthSE": []
                    }
                ],
                "description": "Make a GET Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a GET Request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OauthSE": []
                    }
                ],
                "description": "Make a GET Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a GET Request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OauthSE": []
                    }
                ],
                "description": "Make a GET Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a GET Request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "options": {
                "security": [
                    {
                        "OauthSE": []
                    }
                ],
                "description": "Make a GET Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a GET Request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "OauthSE": []
                    }
                ],
                "description": "Make a GET Request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a GET Request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "OauthSE": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://oauth.se/realms/master/protocol/openid-connect/auth",
            "tokenUrl": "https://oauth.se/realms/master/protocol/openid-connect/token",
            "scopes": {
                "email": "Email Scope",
                "openid": "Default Scope",
                "profile": "Profile Scope"
            }
        }
    }
}